rename flash.js to flash-factory.js
