+++ src/flash-factory.js
